+++ pydis_core/exts/__init__.py
@@ -1,4 +1,8 @@
 """Reusable Discord cogs."""
-__all__ = []
+from pydis_core.exts import source
+
+__all__ = [
+    source,
+]
 
 __all__ = [module.__name__ for module in __all__]